+++ saleor/plugins/webhook/tests/subscription_webhooks/subscription_queries.py
@@ -2001,7 +2001,6 @@ subscription {
              ...on CategoryTranslatableContent {
                id
                name
-               categoryId
              }
           }
         }